Consolidates BuildImageSteps and BuildDockerSteps into BuildSteps.
