+++ jib-core/src/integration-test/java/com/google/cloud/tools/jib/builder/BuildStepsIntegrationTest.java
@@ -19,21 +19,54 @@ package com.google.cloud.tools.jib.builder;
 import com.google.cloud.tools.jib.Command;
 import com.google.cloud.tools.jib.cache.Caches;
 import com.google.cloud.tools.jib.image.ImageReference;
+import com.google.cloud.tools.jib.registry.LocalRegistry;
 import java.nio.file.Path;
 import org.junit.Assert;
+import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-/** Integration tests for {@link BuildDockerSteps}. */
-public class BuildDockerStepsIntegrationTest {
+/** Integration tests for {@link BuildSteps}. */
+public class BuildStepsIntegrationTest {
+
+  @ClassRule public static LocalRegistry localRegistry = new LocalRegistry(5000);
 
   private static final TestBuildLogger logger = new TestBuildLogger();
 
   @Rule public TemporaryFolder temporaryCacheDirectory = new TemporaryFolder();
 
   @Test
-  public void testSteps() throws Exception {
+  public void testSteps_forBuildToDockerRegistry() throws Exception {
+    SourceFilesConfiguration sourceFilesConfiguration = new TestSourceFilesConfiguration();
+    BuildConfiguration buildConfiguration =
+        BuildConfiguration.builder(logger)
+            .setBaseImage(ImageReference.of("gcr.io", "distroless/java", "latest"))
+            .setTargetImage(ImageReference.of("localhost:5000", "testimage", "testtag"))
+            .setMainClass("HelloWorld")
+            .build();
+
+    Path cacheDirectory = temporaryCacheDirectory.newFolder().toPath();
+    BuildSteps buildImageSteps =
+        BuildSteps.forBuildToDockerRegistry(
+            buildConfiguration,
+            sourceFilesConfiguration,
+            Caches.newInitializer(cacheDirectory).setBaseCacheDirectory(cacheDirectory));
+
+    long lastTime = System.nanoTime();
+    buildImageSteps.run();
+    logger.info("Initial build time: " + ((System.nanoTime() - lastTime) / 1_000_000));
+    lastTime = System.nanoTime();
+    buildImageSteps.run();
+    logger.info("Secondary build time: " + ((System.nanoTime() - lastTime) / 1_000_000));
+
+    String imageReference = "localhost:5000/testimage:testtag";
+    new Command("docker", "pull", imageReference).run();
+    Assert.assertEquals("Hello world\n", new Command("docker", "run", imageReference).run());
+  }
+
+  @Test
+  public void testSteps_forBuildToDockerDaemon() throws Exception {
     SourceFilesConfiguration sourceFilesConfiguration = new TestSourceFilesConfiguration();
     BuildConfiguration buildConfiguration =
         BuildConfiguration.builder(logger)
@@ -43,8 +76,8 @@ public class BuildDockerStepsIntegrationTest {
             .build();
 
     Path cacheDirectory = temporaryCacheDirectory.newFolder().toPath();
-    BuildDockerSteps buildDockerSteps =
-        new BuildDockerSteps(
+    BuildSteps buildDockerSteps =
+        BuildSteps.forBuildToDockerDaemon(
             buildConfiguration,
             sourceFilesConfiguration,
             Caches.newInitializer(cacheDirectory).setBaseCacheDirectory(cacheDirectory));
